backend/requirements.txt
# BantuBuzz Backend Requirements
# For quick start with SQLite, use this file
# For PostgreSQL support, use requirements-postgres.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
celery==5.3.4
redis==5.0.1
python-socketio==5.11.0
paynow==1.0.8
email-validator==2.1.0
Pillow==10.1.0
gunicorn==21.2.0
simple-websocket==1.0.0
Werkzeug==3.0.1

# Note: psycopg2-binary removed for easier installation
# If you need PostgreSQL support, run:
# pip install -r requirements-postgres.txt
